+++ src/refactor_nrepl/util.clj
@@ -73,7 +73,12 @@
   (when (System/getProperty "refactor-nrepl.internal.log-exceptions")
     (.printStackTrace e)))
 
-(defn wrap-ignore-errors [f ignore-errors?]
+(defn wrap-ignore-errors
+  "Wraps a predicate `f` in a try-catch, depending on `ignore-errors?`.
+
+  Typically used for making ingestion of data (`read`/`require`/`analyze`) more robust
+  in face of unconfigured or otherwise problematic source paths."
+  [f ignore-errors?]
   (if-not ignore-errors?
     f
     (fn [x]
@@ -81,5 +86,5 @@
         (f x)
         (catch Exception e
           (maybe-log-exception e)
-          ;; return false, because `wrap-ignore-errors` is oriented for predicate usage:
+          ;; return false, because `wrap-ignore-errors` is oriented for predicate usage
           false)))))